Fleet UI: Move show query button in query report
