+++ changes/16701-move-show-query-button
@@ -1,0 +1,1 @@
+- Move show query button so it shows in report page even with no results
