+++ frontend/pages/queries/details/components/QueryReport/QueryReport.tsx
@@ -13,7 +13,6 @@ import { IQueryReport, IQueryReportResultRow } from "interfaces/query_report";
 import Button from "components/buttons/Button";
 import Icon from "components/Icon/Icon";
 import TableContainer from "components/TableContainer";
-import ShowQueryModal from "components/modals/ShowQueryModal";
 import TooltipWrapper from "components/TooltipWrapper";
 import EmptyTable from "components/EmptyTable";
 
@@ -46,7 +45,6 @@ const QueryReport = ({
 }: IQueryReportProps): JSX.Element => {
   const { lastEditedQueryName, lastEditedQueryBody } = useContext(QueryContext);
 
-  const [showQueryModal, setShowQueryModal] = useState(false);
   const [filteredResults, setFilteredResults] = useState<Row[]>(
     flattenResults(queryReport?.results || [])
   );
@@ -77,22 +75,9 @@ const QueryReport = ({
     );
   };
 
-  const onShowQueryModal = () => {
-    setShowQueryModal(!showQueryModal);
-  };
-
   const renderTableButtons = () => {
     return (
       <div className={`${baseClass}__results-cta`}>
-        <Button
-          className={`${baseClass}__show-query-btn`}
-          onClick={onShowQueryModal}
-          variant="text-icon"
-        >
-          <>
-            Show query <Icon name="eye" />
-          </>
-        </Button>
         <Button
           className={`${baseClass}__export-btn`}
           onClick={onExportQueryResults}
@@ -170,17 +155,7 @@ const QueryReport = ({
     );
   };
 
-  return (
-    <div className={`${baseClass}__wrapper`}>
-      {renderTable()}
-      {showQueryModal && (
-        <ShowQueryModal
-          query={lastEditedQueryBody}
-          onCancel={onShowQueryModal}
-        />
-      )}
-    </div>
-  );
+  return <div className={`${baseClass}__wrapper`}>{renderTable()}</div>;
 };
 
 export default QueryReport;
